requirements.txt
langchain==0.3.0
langchain-community
langchain-google-genai
langchain-chroma
chromadb
pypdf
fastapi
uvicorn
mcp
python-dotenv
elevenlabs
unstructured
python-magic-bin
langchain-openai
